+++ lib/services/parameter-store/make-parameter-store.js
@@ -1,31 +1,14 @@
 'use strict';
 
 const Bluebird = require('bluebird');
-const get = require('lodash.get');
-const { sortParameters } = require('./sort-parameters');
-const { makeSsmStore } = require('./stores/ssm/make-ssm-store');
-const { makeDdbStore } = require('./stores/ddb/make-ddb-store');
-
 const chalk = require('chalk');
+
+const { sortParameters } = require('./sort-parameters');
+const { makeGetProviderStore } = require('./make-get-provider-store');
 const { log } = require('../../utils/logger');
 
 const makeParameterStore = ({ settingsService }) => {
-    const getProviderStore = () => {
-      return settingsService.getSettings()
-      .then(settings => {
-        if (get(settings, 'provider.name') === 'ddb') {
-          return makeDdbStore({
-            tableName: get(settings, 'provider.tableName')
-          });
-        }
-
-        if (get(settings, 'provider.name') === 'ssm') {
-          return makeSsmStore({});
-        }
-
-        throw new Error('Unsupported provider specified');
-      });
-    }
+    const getProviderStore = makeGetProviderStore({ settingsService });
 
     const updateSecrets = ({ parameters }) => {
       return getProviderStore()
